Api crud refactoring (#120)
* Refactor and move crud.py to crud folder

* Refactor 'Application' crud to crud_famApplication and fix import for others.

* Refactor crud for User and Application with related imports.

* Refactor test and fixtures for User.

* Crud refactoring for group, user, role

* Crud role refactoring

* Some naming refactoring: remove _fam prefix

* Naming refactoring: remove _fam prefix

* Minor extra space removed.
