+++ backend/api/app/routers/fam_router.py
@@ -1,16 +1,16 @@
 import logging
 from typing import List
 
+from api.app.crud import crud_application, crud_role, crud_user
 from fastapi import APIRouter, Depends, HTTPException, Response
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from .. import crud, dependencies, schemas
+from .. import dependencies, schemas
 
 LOGGER = logging.getLogger(__name__)
 
 router = APIRouter()
-
 
 @router.get(
     "/fam_applications",
@@ -24,11 +24,10 @@ def get_fam_applications(response: Response,
     List of different applications that are administered by FAM
     """
     LOGGER.debug(f"running router ... {db}")
-    queryData = crud.getFamApplications(db)
+    queryData = crud_application.getFamApplications(db)
     if len(queryData) == 0:
         response.status_code = 204
     return queryData
-
 
 @router.post(
     "/fam_applications", response_model=schemas.FamApplication, tags=["FAM_application"]
@@ -41,10 +40,9 @@ def create_fam_application(
     Add Application/client to FAM
     """
     LOGGER.debug(f"running router ... {db}")
-    queryData = crud.createFamApplication(famApplication, db)
+    queryData = crud_application.createFamApplication(famApplication, db)
     # return queryData
     return queryData
-
 
 @router.delete(
     "/fam_applications/{application_id}",
@@ -59,16 +57,15 @@ def delete_fam_application(
     Add Application/client to FAM
     """
     LOGGER.debug(f"running router ... {db}")
-    application = crud.getFamApplication(application_id=application_id, db=db)
+    application = crud_application.getFamApplication(application_id=application_id, db=db)
     if not application:
         raise HTTPException(
             status_code=404, detail=f"application_id={application_id} does not exist"
         )
     application_id = application.application_id
     LOGGER.debug(f"application_id: {application_id}")
-    application = crud.deleteFamApplication(db=db, application_id=application_id)
+    application = crud_application.deleteFamApplication(db=db, application_id=application_id)
     return application
-
 
 @router.get("/fam_users", response_model=List[schemas.FamUserGet], tags=["FAM_users"])
 def get_fam_users(db: Session = Depends(dependencies.get_db)):
@@ -76,9 +73,8 @@ def get_fam_users(db: Session = Depends(dependencies.get_db)):
     List of different users that are administered by FAM
     """
     LOGGER.debug(f"running router ... {db}")
-    queryData = crud.getFamUsers(db)
+    queryData = crud_user.getFamUsers(db)
     return queryData
-
 
 @router.post("/fam_users", response_model=schemas.FamUserGet, tags=["FAM_users"])
 def create_fam_user(
@@ -90,7 +86,7 @@ def create_fam_user(
     queryData = None
     LOGGER.debug(f"running router ... {db}")
     try:
-        queryData = crud.createFamUser(famUser, db)
+        queryData = crud_user.createFamUser(famUser, db)
         LOGGER.debug(f"queryData: {queryData}")
     except IntegrityError as e:
         LOGGER.debug(f"error: {e}")
@@ -101,7 +97,6 @@ def create_fam_user(
 
     return queryData
 
-
 @router.delete(
     "/fam_users/{user_id}", response_model=schemas.FamUser, tags=["FAM_users"]
 )
@@ -110,13 +105,12 @@ def delete_fam_user(user_id: int, db: Session = Depends(dependencies.get_db)):
     Delete a FAM user
     """
 
-    user = crud.getFamUser(user_id=user_id, db=db)
+    user = crud_user.getFamUser(user_id=user_id, db=db)
     LOGGER.debug(f"user: {user}")
     if not user:
         raise HTTPException(status_code=404, detail=f"user_id={user_id} does not exist")
-    user = crud.deleteUser(db=db, user_id=user_id)
+    user = crud_user.deleteUser(db=db, user_id=user_id)
     return user
-
 
 @router.get(
     "/fam_users/{user_id}", response_model=schemas.FamUserGet, tags=["FAM_users"]
@@ -126,10 +120,9 @@ def get_fam_user(user_id: int, db: Session = Depends(dependencies.get_db)):
     Get a FAM user
     """
     LOGGER.debug(f"userid is: {user_id}")
-    user = crud.getFamUser(user_id=user_id, db=db)
+    user = crud_user.getFamUser(user_id=user_id, db=db)
     LOGGER.debug(f"userdata: {user}")
     return user
-
 
 @router.get("/fam_roles",
             response_model=List[schemas.FamRoleGet],
@@ -139,9 +132,8 @@ def get_fam_roles(db: Session = Depends(dependencies.get_db)):
     List of different roles that are administered by FAM
     """
     LOGGER.debug(f"running router ... {db}")
-    queryData = crud.getFamRoles(db)
+    queryData = crud_role.getFamRoles(db)
     return queryData
-
 
 @router.post("/fam_roles",
              response_model=schemas.FamRoleGet,
@@ -155,7 +147,7 @@ def create_fam_role(
     queryData = None
     LOGGER.debug(f"running router ... {db}")
     try:
-        queryData = crud.createFamRole(famRole, db)
+        queryData = crud_role.createFamRole(famRole, db)
         LOGGER.debug(f"queryData: {queryData}")
     except IntegrityError as e:
         LOGGER.debug(f"error: {e}")
