+++ backend/tests/conftest.py
@@ -16,16 +16,12 @@ import sys
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 # sys.path.append(os.path.dirname(__file__))
 
-import datetime
 import logging
 import sys
-import uuid
-from typing import Any, Generator, TypedDict
+from typing import Any, Generator
 
-import api.app.crud as crud
 import api.app.dependencies as dependencies
 import api.app.models.model as model
-import api.app.schemas as schemas
 import pytest
 from api.app.database import Base
 from api.app.main import app
@@ -33,7 +29,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from sqlalchemy import create_engine, event
 from sqlalchemy.engine.base import Engine
-from sqlalchemy.orm import session, sessionmaker
+from sqlalchemy.orm import sessionmaker
 
 # global placeholder to be populated by fixtures for database test
 # sessions, required to override the get_db method.
@@ -46,21 +42,9 @@ LOGGER = logging.getLogger(__name__)
 pytest_plugins = [
     "fixtures.fixtures_crud_application",
     "fixtures.fixtures_router_application",
+    "fixtures.fixtures_crud_user",
+    "fixtures.fixtures_crud_role"
 ]
-
-
-class FamUserTD(TypedDict):
-    # cludge... ideally this type should be derived from the
-    # pydantic model schema.FamUser
-    user_type: str
-    cognito_user_id: str
-    user_name: str
-    user_guid: str
-    create_user: str
-    create_date: datetime.datetime
-    update_user: str
-    update_date: datetime.datetime
-
 
 @pytest.fixture(scope="function")
 def getApp(sessionObjects, dbEngine: Engine) -> Generator[FastAPI, Any, None]:
@@ -73,7 +57,6 @@ def getApp(sessionObjects, dbEngine: Engine) -> Generator[FastAPI, Any, None]:
     app.dependency_overrides[dependencies.get_db] = override_get_db
     yield app
 
-
 # This @event is important. By default FOREIGN KEY constraints have no effect on the operation of the table from SQLite.
 # It (FOREIGN KEY) only works when emitting CREATE statements for tables.
 # Reference: https://docs.sqlalchemy.org/en/14/dialects/sqlite.html#foreign-key-support
@@ -82,7 +65,6 @@ def set_sqlite_pragma(dbapi_connection, connection_record):
     cursor = dbapi_connection.cursor()
     cursor.execute("PRAGMA foreign_keys=ON")
     cursor.close()
-
 
 @pytest.fixture(scope="function")
 def testClient_fixture(getApp: FastAPI) -> TestClient:
@@ -97,14 +79,12 @@ def testClient_fixture(getApp: FastAPI) -> TestClient:
     client = TestClient(getApp)
     yield client
 
-
 @pytest.fixture(scope="module")
 def sessionObjects(dbEngine: Engine) -> sessionmaker:
     # Use connect_args parameter only with sqlite
     SessionTesting = sessionmaker(autocommit=False, autoflush=False, bind=dbEngine)
     LOGGER.debug(f"session type: {type(SessionTesting)}")
     yield SessionTesting
-
 
 @pytest.fixture(scope="module")
 def dbEngine() -> Engine:
@@ -137,7 +117,6 @@ def dbEngine() -> Engine:
         LOGGER.debug("remove the database: ./test_db.db'")
         os.remove("./test_db.db")
 
-
 @pytest.fixture(scope="function")
 def dbSession(dbEngine, sessionObjects) -> Generator[sessionObjects, Any, None]:
 
@@ -149,231 +128,6 @@ def dbSession(dbEngine, sessionObjects) -> Generator[sessionObjects, Any, None]:
     session.close()
     # transaction.rollback()
     connection.close()
-
-
-@pytest.fixture(scope="function")
-def dbSession_famUsers_withdata(
-    dbSession, testUserData3, testGroupData, userGroupXrefData
-):
-    """to add a user need to satisfy the integrity constraints:
-
-    1. create the group
-    2. retrieve the group id
-    3.
-
-    :param dbSession: _description_
-    :type dbSession: _type_
-    :param testUserData: _description_
-    :type testUserData: _type_
-    :param add_group: _description_
-    :type add_group: _type_
-    :yield: _description_
-    :rtype: _type_
-    """
-    # the following link goes over working with related/associated tables
-    # https://www.pythoncentral.io/sqlalchemy-association-tables/
-
-    db = dbSession
-    # trying to add to user without violating the integrity constraint
-    # group was populated with a record by the add_group fixture.
-    newUser = model.FamUser(**testUserData3)
-    groupSchema = model.FamGroup(**testGroupData)
-
-    userGroupXrefData["group"] = groupSchema
-    userGroupXrefData["user"] = newUser
-
-    xrefTable = model.FamUserGroupXref(**userGroupXrefData)
-    db.add(xrefTable)
-    db.commit()
-
-    yield db
-
-    db.delete(xrefTable)
-    db.delete(groupSchema)
-    db.delete(newUser)
-
-    db.commit()
-
-
-@pytest.fixture(scope="function")
-def userGroupXrefData():
-    nowdatetime = datetime.datetime.now()
-    xrefData = {
-        "create_user": "serg",
-        "create_date": nowdatetime,
-        "update_user": "ron",
-        "update_date": nowdatetime,
-    }
-    yield xrefData
-
-
-@pytest.fixture(scope="function")
-def add_group(dbSession, testGroupData):
-    db = dbSession
-    groupSchema = schemas.FamGroupPost(**testGroupData)
-
-    crud.createFamGroup(famGroup=groupSchema, db=db)
-    yield db
-
-    db.delete(testGroupData)
-    db.commit()
-
-
-@pytest.fixture(scope="function")
-def testGroupData():
-    testGroupData = {
-        "group_name": "test group",
-        "purpose": "testing",
-        "create_user": "Brian Trotier",
-        "create_date": datetime.datetime.now()
-    }
-    return testGroupData
-
-
-@pytest.fixture(scope="function")
-def testUserData_asPydantic(testUserData) -> schemas.FamUser:
-    famUserAsPydantic = schemas.FamUser(**testUserData)
-    yield famUserAsPydantic
-
-
-@pytest.fixture(scope="function")
-def testUserData2_asPydantic(testUserData2) -> schemas.FamUser:
-    famUserAsPydantic2 = schemas.FamUser(**testUserData2)
-    yield famUserAsPydantic2
-
-
-@pytest.fixture(scope="function")
-def deleteAllUsers(dbSession: session.Session) -> None:
-    """Cleans up all users from the database after the test has been run
-
-    :param dbSession: mocked up database session
-    :type dbSession: sqlalchemy.orm.session.Session
-    """
-    LOGGER.debug(f"dbsession type: {type(dbSession)}")
-    yield
-    db = dbSession
-    famUsers = db.query(model.FamUser).all()
-    for famUser in famUsers:
-        db.delete(famUser)
-    db.commit()
-
-
-@pytest.fixture(scope="function")
-def testUserData() -> dict:
-
-    userData = {
-        "user_type": "a",
-        "cognito_user_id": "22ftw",
-        "user_name": "Mike Bossy",
-        "user_guid": str(uuid.uuid4()),
-        "create_user": "Al Arbour",
-        "create_date": datetime.datetime.now(),
-        "update_user": "Al Arbour",
-        "update_date": datetime.datetime.now(),
-    }
-    yield userData
-
-
-@pytest.fixture(scope="function")
-def testUserData2() -> FamUserTD:
-    userData = {
-        "user_type": "a",
-        "cognito_user_id": "22dfs",
-        "user_name": "Dennis Potvin",
-        "user_guid": str(uuid.uuid4()),
-        "create_user": "Al Arbour",
-        "create_date": datetime.datetime.now(),
-        "update_user": "Al Arbour",
-        "update_date": datetime.datetime.now(),
-    }
-    yield userData
-
-
-@pytest.fixture(scope="function")
-def dbSession_famRoles_withSimpleData(dbSession, simpleRoleData):
-    db = dbSession
-    # add a record to the database
-    newRole = model.FamRole(**simpleRoleData)
-    db.add(newRole)
-    db.commit()
-    yield db  # use the session in tests.
-
-    db.delete(newRole)
-    db.commit()
-
-
-@pytest.fixture(scope="function")
-def testUserData3() -> FamUserTD:
-    userData = {
-        "user_type": "a",
-        "cognito_user_id": "zzff",
-        "user_name": "Billy Smith",
-        "user_guid": str(uuid.uuid4()),
-        "create_user": "Al Arbour"
-    }
-    yield userData
-
-
-@pytest.fixture(scope="function")
-def addApplication(dbSession, testApplicationData):
-    """ This test is going to add and application record to the database and
-    then when the test is torn down it will remove that record.
-
-    Args:
-        dbSession (_type_): _description_
-        testApplicationData (_type_): _description_
-    """
-    db = dbSession
-    newApp = model.FamApplication(**testApplicationData)
-    db.add(newApp)
-    db.commit()
-    yield db  # use the session in tests.
-
-    db.delete(newApp)
-    db.commit()
-
-
-@pytest.fixture(scope="function")
-def testApplicationData(dbSession):
-    appData = {
-        "application_name" : 'testapp',
-        "applicationdescription": "test description"
-
-    }
-    yield appData
-
-
-@pytest.fixture(scope="function")
-def simpleRoleData_asPydantic(simpleRoleData) -> schemas.FamRole:
-    famRoleAsPydantic = schemas.FamRole(**simpleRoleData)
-    yield famRoleAsPydantic
-
-
-@pytest.fixture(scope="function")
-def simpleRoleData() -> dict:
-    roleData = {
-        "role_name": "FAM_ADMIN",
-        "role_purpose": "FAM Admin",
-        "create_user": "John Doe"
-    }
-    yield roleData
-
-
-@pytest.fixture(scope="function")
-def deleteAllRoles(dbSession: session.Session) -> None:
-    """Cleans up all roles from the database after the test has been run
-
-    :param dbSession: mocked up database session
-    :type dbSession: sqlalchemy.orm.session.Session
-    """
-    LOGGER.debug(f"dbsession type: {type(dbSession)}")
-    yield
-    db = dbSession
-    famRoles = db.query(model.FamRole).all()
-    for famRole in famRoles:
-        db.delete(famRole)
-    db.commit()
-
 
 def override_get_db():
     try:
