+++ backend/tests/fixtures/fixtures_crud_application.py
@@ -4,7 +4,7 @@ import pytest
 
 import api.app.models.model as model
 import api.app.schemas as schemas
-import api.app.crud as crud
+from api.app.crud import crud_application as crud_application
 
 LOGGER = logging.getLogger(__name__)
 
@@ -12,7 +12,7 @@ LOGGER = logging.getLogger(__name__)
 @pytest.fixture(scope="function")
 def dbSession_famApplication_withdata(dbSession, applicationData1):
     applicationData1AsPydantic = schemas.FamApplicationCreate(**applicationData1)
-    appData = crud.createFamApplication(famApplication=applicationData1AsPydantic,
+    appData = crud_application.createFamApplication(famApplication=applicationData1AsPydantic,
                                         db=dbSession)
 
     yield dbSession
